This will separate manual_tests
